+++ apps/playground/src/ts/index.ts
@@ -6,10 +6,12 @@ import "./speculate/speculator";
 import loopsStore from "./elements/loops";
 import exampleStore from "./elements/example";
 import navStore from "./elements/nav";
+import bindsStore from "./elements/binds";
 
 storeModule("loops", loopsStore);
 storeModule("example", exampleStore);
 storeModule("nav", navStore);
+storeModule("binds", bindsStore);
 
 Elements.start({
 	debug: true,